scroll the viewport based on the player's position
Set a margin from the edge of the viewport in GameMap, uses adjust_viewport_anchor() to move the viewport to maintain this margin at the start of render(), and do a bunch of fiddly coordinate math to calculate the part of the map that's inside the viewport and what region of the viewport the map should render into.

Define a new tile type, EXTERNAL, for tiles that are in the viewport but are not in the map. This allows the viewport to scroll beyond the edge of the map if the player forces this by moving close to the edge of the map. EXTERNAL is visually identical to SHROUD for lack of any better idea.

GameMap.render() now performs a very janky quasi-double buffering. To avoid reading beyond the bounds of the GameMap's tiles array render() creates a viewport-sized array of EXTERNAL before blitting in the relevant portion of the map tiles.